tulisanai-backend/requirements.txt
flask
pillow
pytesseract
kivy
plyer
requests